requirements.txt
asgiref==3.11.0
Django==5.2.8
django-nested-admin==4.1.6
pillow==12.0.0
python-monkey-business==1.1.0
sqlparse==0.5.4
tzdata==2025.2
